Remove full layout. Let controllers decide when the layout is full or not.
There's less duplication having just one layout and letting the controller
actions decide when to use or not the sidebar. This adds coupling between
layout & actions and adds a little bit of logic to the layout, but it's
better than having to change 2 files to make a global layout change.

Most actions should just use the non-full layout (by default we have a
sidebar), so this will not be a big problem.