Don't die when GitHub won't give me all the information for a user.
For now, let's just require the username, but this is getting fucked up
now.

Ideally we'll just get the information from GitHub and show a form to
the user so they can fill it up automatically after it's all done.

This way we can have email confirmation to activate accounts and all
those sort of things and not depend on the GitHub API to gather the
required information.